Projects/Social-Media-Scraping/insta_scraper_02/requirements.txt
playwright
python-dotenv
pandas
requests
openpyxl
Pillow
transformers